msa_framework/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
PyJWT==2.10.1
python-dotenv==1.0.1

